refactor(app): extract record creation logic into a separate module
