+++ src/app/page.tsx
@@ -6,21 +6,14 @@ import { zodResolver } from '@hookform/resolvers/zod'
 import { InputSchema, InputSchemaDefaultValue } from '@/models/input.dto'
 import { Form } from '@/components/ui/form'
 import { Button } from '@/components/ui/button'
-import {
-  detectRecordFormat,
-  importCSA,
-  importJKFString,
-  importKI2,
-  importKIF,
-  Position,
-  Record,
-  RecordFormatType
-} from 'tsshogi'
 import { client } from '@/lib/client'
 import { useState } from 'react'
 import { AxiosError } from 'axios'
 import { toast } from 'sonner'
 import type { ResultSchema } from '@/models/result.dto'
+import { createRecord } from '@/lib/record'
+import type { Record } from 'tsshogi'
+import { toSHA256Hash } from '@/lib/hash'
 
 export default function Page() {
   const form = useForm<InputSchema>({
@@ -29,90 +22,11 @@ export default function Page() {
   })
   const { control } = form
   const [_result, setResult] = useState<ResultSchema | null>(null)
-  const toSHA256Hash = async (record: Record): Promise<string> => {
-    const encoder = new TextEncoder()
-    const data = encoder.encode(record.initialPosition.sfen)
-    const buffer = await crypto.subtle.digest('SHA-256', data)
-    return Array.from(new Uint8Array(buffer))
-      .map((b) => b.toString(16).padStart(2, '0'))
-      .join('')
-  }
-
-  const handleRecord = (data: InputSchema): Record => {
-    const type: RecordFormatType = detectRecordFormat(data.content)
-    console.log('Detected record format:', type)
-    switch (type) {
-      case RecordFormatType.KIF: {
-        const _record: Record | Error = importKIF(data.content)
-        console.log('Imported KIF record:', _record)
-        if (_record instanceof Error) {
-          console.error('Invalid KIF format:', _record.message)
-          throw _record
-        }
-        return _record
-      }
-      case RecordFormatType.KI2: {
-        const _record: Record | Error = importKI2(data.content)
-        if (_record instanceof Error) {
-          console.error('Invalid KI2 format:', _record.message)
-          throw _record
-        }
-        return _record
-      }
-      case RecordFormatType.CSA: {
-        const _record: Record | Error = importCSA(data.content)
-        if (_record instanceof Error) {
-          console.error('Invalid CSA format:', _record.message)
-          throw _record
-        }
-        return _record
-      }
-      case RecordFormatType.SFEN: {
-        const _position: Position | null = Position.newBySFEN(data.content)
-        if (_position === null) {
-          console.error('Invalid SFEN format:', data.content)
-          throw new Error('Invalid SFEN format')
-        }
-        const _record: Record | Error = new Record(_position)
-        if (_record instanceof Error) {
-          console.error('Invalid SFEN format:', _record.message)
-          throw _record
-        }
-        return _record
-      }
-      case RecordFormatType.USI: {
-        const _record: Record | Error = Record.newByUSI(data.content)
-        if (_record instanceof Error) {
-          console.error('Invalid SFEN format:', _record.message)
-          throw _record
-        }
-        return _record
-      }
-      case RecordFormatType.JKF: {
-        const _record: Record | Error = importJKFString(data.content)
-        if (_record instanceof Error) {
-          console.error('Invalid SFEN format:', _record.message)
-          throw _record
-        }
-        return _record
-      }
-      case RecordFormatType.USEN: {
-        const _record: Record | Error = Record.newByUSEN(data.content)
-        if (_record instanceof Error) {
-          console.error('Invalid SFEN format:', _record.message)
-          throw _record
-        }
-        return _record
-      }
-    }
-  }
 
   const onSubmit = async (data: InputSchema) => {
     try {
-      const record: Record = handleRecord(data)
-      console.log('Record:', record)
+      const record: Record = createRecord(data)
       const hash: string = await toSHA256Hash(record)
-      console.log('Hash:', hash)
       setResult(await client.post('/hash', { hash }))
     } catch (error) {
       console.error('Error fetching result:', error)
@@ -132,10 +46,7 @@ export default function Page() {
     <div className='flex flex-col max-w-xl w-full mx-auto p-6 gap-4'>
       <h1 className='text-center text-2xl font-bold'>詰将棋同一検索ページ</h1>
       <div className='flex flex-col gap-2'>
-        <p className='text-sm'>
-          標準的な棋譜形式(KIF, KI2, CSA, SFEN/USI,
-          JKF)に対応しています。フォームに貼り付けることで完全一致の作品があるかどうかをチェックします。
-        </p>
+        <p className='text-sm'>標準的な棋譜形式(KIF, KI2, CSA, SFEN/USI, JKF)に対応しています。</p>
         <p className='text-sm'>
           同一作のチェックは初期配置(盤面+持ち駒)を利用しています。受け方の持ち駒に制約がある場合、正しく判定できない可能性があります。
         </p>
